pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Simple page analyze program for learning Flask and web dev"
authors = ["Igor Loskutov <omfi017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.3"
flake8 = "^6.1.0"
gunicorn = "^21.2.0"
psycopg = "^3.1.12"
python-dotenv = "^1.0.0"
validators = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
